deps: remove `reth-evm` as a dependency of `precompiles` (#384)
This dependency was mostly used for `reth_evm::revm`, but we can pull in
`revm` directly to make the crate graph leaner. This matters for an
eventual Foundry integration where we would not want to depend on
a lot of Reth internals.

The solution is to use `revm` directly instead of
the `revm` re-exports in `reth_evm`, and `alloy_evm` instead of the
`alloy_evm` re-exports in `reth_evm`.

Some of the reth internals still remain in the crate graph for
`precompiles`, but this is due to `reth-storage-api`. For a leaner
Foundry integration, we should consider feature gating this.

~On top of #383 for now, will rebase once merged.~